refactor complex function into separate module
